refactor(aigateway): split the oversized test modules into behaviour siblings
The seven test modules this branch authored above the 450-line limit each split
along a seam the file already marked, so the halves are two subjects rather than
two arbitrary halves. Every test function, its decorator, its fixtures and its
local helpers moved verbatim; nothing was renamed, re-parametrized, merged or
reworded, and no production source file changed.

The append-only gate certifies a source split by proving no test file changed at
all, which a test relocation makes red by construction. It is replaced here by
the collected node-ID multiset, compared across the whole suite with the module
path normalized away: 2114 before, 2114 after, identical. A test lost,
duplicated, renamed, re-parametrized or newly skipped in the move would appear
in that comparison. It is a multiset because 32 test-name suffixes live in more
than one file, and a set comparison would hide a loss.

Harnesses were copied rather than shared wherever a copy is what preserves
behaviour: _api_key_validation_ok is autouse, and an autouse fixture applies only
to the module that declares it, so importing it would be a different thing and a
conftest would widen it to every sibling. The one shared module,
_openapi_document, holds a slice of the real OpenRouter document whose whole
value is that it was not invented; two copies could drift into two different
documents. No test mutates it and its client holds no class-level state.

Refs: OME-654
