+++ apps/aigateway/tests/unit/openrouter/test_openrouter_openapi_endpoint_source.py
@@ -1,4 +1,4 @@
-"""OME-647 (OME-479 §5.1/§6.1): the OpenAPI endpoint source, parsed and WIRED.
+"""OME-647 (OME-479 §5.1/§6.1): the OpenAPI endpoint source, parsed.
 
 FEATURE: OpenRouter's source PAIR. §5.1 names two fixed public documents — the
 per-model catalog for what a model supports, and the public OpenAPI document for
@@ -16,19 +16,20 @@ enables a parameter, moves gateway.status, changes the /v1/models summary, or
 touches dispatch — only a rule does.
 INVARIANT (§5.2): every fetch goes through the bounded transport under an INJECTED
 client. No test in this file reaches the network.
+
+AIDEV-NOTE: the production wiring lives in ``test_openrouter_openapi_endpoint_route``
+— a parser that is correct against a fixture and never reached by the route is the
+exact failure this pair of files exists to separate.
 """
 
 from __future__ import annotations
 
 import json
-from typing import Any, cast
+from typing import Any
 
 import pytest
-from fastapi import FastAPI
-from fastapi.testclient import TestClient
 
 from aigateway.core.chat_parameters import ProviderParameterObservation, overlay_observations
-from aigateway.core.discovery_runtime import DiscoveryRuntime
 from aigateway.core.parameter_discovery import (
     DiscoveryError,
     DiscoveryHttpClient,
@@ -36,9 +37,6 @@ from aigateway.core.parameter_discovery import (
     RawResponse,
     fetch_discovery_json,
 )
-from aigateway.core.parameter_discovery_cache import CacheLimits, ObservationCache
-from aigateway.core.profile_index import ProfileIndexStore
-from aigateway.core.profile_models import Profile, ProfileState, profile_id_for
 from aigateway.plugins.openrouter_provider.discovery import (
     ALLOWED_ORIGINS,
     CHAT_REQUEST_SCHEMA,
@@ -47,90 +45,8 @@ from aigateway.plugins.openrouter_provider.discovery import (
     openapi_discovery_limits,
     parse_openapi_endpoint_observations,
 )
-from aigateway.plugins.openrouter_provider.settings import OpenRouterPluginSettings
 
-_UPSTREAM = "google/gemini-2.0-flash-001"
-_MODEL = f"openrouter/{_UPSTREAM}"
-
-# Slice of the VERIFIED live document (fetched and measured 2026-07-28). Every shape
-# below is copied from it, not invented: the ``["number", "null"]`` type unions, the
-# prose-only ranges, the ``anyOf`` on stop, the enum-plus-null on service_tier, and
-# `route`'s deprecation hidden one `$ref` hop away in ``DeprecatedRoute``.
-_OPENAPI: dict[str, Any] = {
-    "openapi": "3.1.0",
-    "components": {
-        "schemas": {
-            CHAT_REQUEST_SCHEMA: {
-                "type": "object",
-                "required": ["messages"],
-                "properties": {
-                    "model": {"$ref": "#/components/schemas/ModelName"},
-                    "messages": {"type": "array"},
-                    "stream": {"type": "boolean", "default": False},
-                    "temperature": {
-                        "description": "Sampling temperature (0-2)",
-                        "format": "double",
-                        "type": ["number", "null"],
-                    },
-                    "top_k": {"type": ["integer", "null"]},
-                    "seed": {"type": ["integer", "null"]},
-                    "stop": {
-                        "anyOf": [
-                            {"type": "string"},
-                            {"items": {"type": "string"}, "maxItems": 4, "type": "array"},
-                            {"type": "null"},
-                        ],
-                        "description": "Stop sequences (up to 4)",
-                    },
-                    "service_tier": {
-                        "enum": ["auto", "default", "flex", None],
-                        "type": ["string", "null"],
-                    },
-                    "tool_choice": {"$ref": "#/components/schemas/ChatToolChoice"},
-                    "response_format": {
-                        "description": "Response format configuration",
-                        "discriminator": {"propertyName": "type"},
-                        "oneOf": [{"$ref": "#/components/schemas/ChatFormatJsonObject"}],
-                    },
-                    "route": {"$ref": "#/components/schemas/DeprecatedRoute"},
-                    "max_tokens": {
-                        "description": "Maximum tokens (deprecated, use max_completion_tokens).",
-                        "type": ["integer", "null"],
-                    },
-                },
-            },
-            "ModelName": {"type": "string"},
-            "DeprecatedRoute": {
-                "deprecated": True,
-                "enum": ["fallback", "sort", None],
-                "type": ["string", "null"],
-                "x-speakeasy-deprecation-message": "Use providers.sort.partition instead",
-            },
-            "ChatToolChoice": {
-                "anyOf": [
-                    {"enum": ["none"], "type": "string"},
-                    {"enum": ["auto"], "type": "string"},
-                    {"$ref": "#/components/schemas/ChatNamedToolChoice"},
-                ]
-            },
-            "ChatNamedToolChoice": {"type": "object"},
-            "ChatFormatJsonObject": {"type": "object"},
-        }
-    },
-}
-
-_CATALOG = {
-    "data": [
-        {
-            "id": _UPSTREAM,
-            "supported_parameters": ["temperature", "max_tokens", "seed", "top_k"],
-        }
-    ]
-}
-
-
-def _by_path(obs: tuple[ProviderParameterObservation, ...]) -> dict[str, Any]:
-    return {o.request_path: o for o in obs}
+from ._openapi_document import _CATALOG, _OPENAPI, _UPSTREAM, _by_path, _RoutingClient
 
 
 def _endpoint() -> dict[str, Any]:
@@ -376,27 +292,6 @@ async def test_the_catalog_fetch_keeps_the_operators_bounds() -> None:
 # --- the snapshot ------------------------------------------------------------
 
 
-class _RoutingClient(DiscoveryHttpClient):
-    """Canned JSON per URL; records the URL and bounds of every dial."""
-
-    def __init__(self, bodies: dict[str, Any], *, fail: str | None = None) -> None:
-        self._bodies = bodies
-        self._fail = fail
-        self.seen: list[tuple[str, float, int]] = []
-
-    @property
-    def calls(self) -> list[str]:
-        return [url for url, _timeout, _max_bytes in self.seen]
-
-    async def get(self, url: str, *, timeout_s: float, max_bytes: int) -> RawResponse:
-        self.seen.append((url, timeout_s, max_bytes))
-        if self._fail == url:
-            raise DiscoveryError("unreachable")
-        return RawResponse(
-            status=200, content_type="application/json", body=json.dumps(self._bodies[url])
-        )
-
-
 @pytest.mark.asyncio
 async def test_the_snapshot_now_carries_endpoint_evidence_from_the_openapi_document() -> None:
     from aigateway.plugins.openrouter_provider.discovery import discover_openrouter_snapshot
@@ -429,177 +324,3 @@ async def test_only_the_two_fixed_public_documents_are_ever_dialed() -> None:
     client = _RoutingClient({MODELS_URL: _CATALOG, OPENAPI_URL: _OPENAPI})
     await discover_openrouter_snapshot(_UPSTREAM, client=client)
     assert client.calls == [MODELS_URL, OPENAPI_URL]  # no retry storm, no third host
-
-
-# --- production wiring: the route, not a fixture -----------------------------
-#
-# WHY this section exists: `parse_openapi_endpoint_observations` returns () for a
-# schema name it cannot find. A wiring that used the wrong component name would
-# raise nothing, log nothing, and pass every fixture test above while publishing an
-# empty endpoint source. Only a test that walks the real route can catch that.
-
-
-@pytest.fixture
-def openrouter_enabled(monkeypatch: pytest.MonkeyPatch) -> None:
-    # patches the singleton INSTANCE, not the environment: `load_plugins` hands the
-    # same object to every app, so env vars set after import cannot reach it.
-    from aigateway.plugins.openrouter_provider import plugin as plugin_module
-
-    monkeypatch.setattr(
-        plugin_module.PLUGIN,
-        "settings",
-        OpenRouterPluginSettings(enabled=True, default_models=[_MODEL]),
-    )
-
-
-def _install_runtime(client: TestClient, http: DiscoveryHttpClient) -> None:
-    app = cast(FastAPI, client.app)
-    app.state.discovery_runtime = DiscoveryRuntime(
-        client=http,
-        cache=ObservationCache(
-            clock=_Clock(), limits=CacheLimits(ttl_s=60.0, stale_ttl_s=120.0, max_entries=8)
-        ),
-        limits=DiscoveryLimits(),
-    )
-
-
-class _Clock:
-    def now(self) -> float:
-        return 1000.0
-
-
-async def _contract(credential_blobs, client: TestClient) -> dict[str, Any]:
-    account_id = client.get("/v1/auth/me").json()["id"]
-    await ProfileIndexStore(credential_store=credential_blobs.store).upsert(
-        Profile(
-            id=profile_id_for(account_id, "openrouter", "default"),
-            account_id=account_id,
-            provider="openrouter",
-            name="default",
-            state=ProfileState.AUTHENTICATED,
-            auth_type="api_key",
-        )
-    )
-    resp = client.get("/v1/model-parameters", params={"model": _MODEL})
-    assert resp.status_code == 200, resp.text
-    return resp.json()
-
-
-@pytest.mark.asyncio
-async def test_the_route_publishes_evidence_sourced_from_the_openapi_document(
-    openrouter_enabled, authenticated_client, credential_blobs
-) -> None:
-    _install_runtime(
-        authenticated_client, _RoutingClient({MODELS_URL: _CATALOG, OPENAPI_URL: _OPENAPI})
-    )
-    body = await _contract(credential_blobs, authenticated_client)
-    sources = {row["provider"]["source"] for row in body["parameters"].values()}
-    assert "openrouter:openapi" in sources
-
-
-@pytest.mark.asyncio
-async def test_the_route_publishes_the_declared_shape_of_an_unprojected_field(
-    openrouter_enabled, authenticated_client, credential_blobs
-) -> None:
-    # service_tier has no gateway rule, so this row is visible-but-DISABLED — and the
-    # endpoint schema is the only shape a client can see for it.
-    _install_runtime(
-        authenticated_client, _RoutingClient({MODELS_URL: _CATALOG, OPENAPI_URL: _OPENAPI})
-    )
-    row = (await _contract(credential_blobs, authenticated_client))["parameters"]["service_tier"]
-    assert row["gateway"]["status"] == "disabled"
-    assert row["gateway"]["reason"] == "projection_not_implemented"
-    assert row["schema"] == {"type": "string", "enum": ["auto", "default", "flex"]}
-
-
-@pytest.mark.asyncio
-async def test_the_route_publishes_the_providers_deprecation_verdict(
-    openrouter_enabled, authenticated_client, credential_blobs
-) -> None:
-    _install_runtime(
-        authenticated_client, _RoutingClient({MODELS_URL: _CATALOG, OPENAPI_URL: _OPENAPI})
-    )
-    params = (await _contract(credential_blobs, authenticated_client))["parameters"]
-    assert params["route"]["provider"]["deprecated"] is True
-    assert params["temperature"]["provider"]["deprecated"] is False
-
-
-@pytest.mark.asyncio
-async def test_a_gateway_owned_rule_schema_still_outranks_the_endpoints(
-    openrouter_enabled, authenticated_client, credential_blobs
-) -> None:
-    # the endpoint declares a bare `{"type": "number"}` for temperature; the gateway's
-    # own rule carries the bounds it VALIDATES against. Evidence must never displace
-    # the schema the gateway actually enforces.
-    _install_runtime(
-        authenticated_client, _RoutingClient({MODELS_URL: _CATALOG, OPENAPI_URL: _OPENAPI})
-    )
-    row = (await _contract(credential_blobs, authenticated_client))["parameters"]["temperature"]
-    assert row["gateway"]["status"] == "enabled"
-    assert row["schema"]["maximum"] is not None
-
-
-def test_the_source_revision_and_lifecycle_both_reach_contract_identity() -> None:
-    """Reading DIFFERENT documents must move the opaque ids, even byte-for-byte.
-
-    Two distinct inputs, one property: a contract_id that did not move would tell a
-    client "nothing changed" while the evidence underneath it came from a different
-    source, or carried a lifecycle verdict it did not carry before.
-    """
-    from aigateway.core.model_parameter_contract import build_model_parameter_document
-
-    def _document(*, source_revision: str, deprecated: bool | None) -> dict[str, Any]:
-        return build_model_parameter_document(
-            canonical_id=_MODEL,
-            gateway_provider="openrouter",
-            auth_mode="api_key",
-            scope="account_profile",
-            context_identity="acct:test|prof:1",
-            rules=(),
-            observations=(
-                ProviderParameterObservation(
-                    request_path="route",
-                    support="supported",
-                    source="openrouter:openapi",
-                    deprecated=deprecated,
-                ),
-            ),
-            tools=(),
-            transport=(),
-            freshness={"stale": False, "degraded": False},
-            source_revision=source_revision,
-        )
-
-    baseline = _document(source_revision="rev-a", deprecated=False)
-
-    # 1. the source pair changed; the observations are byte-identical.
-    moved_source = _document(source_revision="rev-b", deprecated=False)
-    assert moved_source["contract_id"] != baseline["contract_id"]
-    assert moved_source["context"]["revision"] != baseline["context"]["revision"]
-
-    # 2. the lifecycle verdict changed; everything else is identical. Both
-    #    transitions count — including the one OUT of silence, which is why the
-    #    tri-state is encoded distinctly rather than collapsed to a bool.
-    for verdict in (True, None):
-        moved_lifecycle = _document(source_revision="rev-a", deprecated=verdict)
-        assert moved_lifecycle["contract_id"] != baseline["contract_id"], verdict
-        assert moved_lifecycle["context"]["revision"] != baseline["context"]["revision"], verdict
-
-
-@pytest.mark.asyncio
-async def test_the_new_source_moves_no_gateway_decision(
-    openrouter_enabled, authenticated_client, credential_blobs
-) -> None:
-    # F12 semantics, owner-locked: dynamic observations move the EVIDENCE axis only.
-    # A deprecated, endpoint-observed `route` must not become dispatchable, and the
-    # summary must not learn anything from a document the gateway merely read.
-    _install_runtime(
-        authenticated_client, _RoutingClient({MODELS_URL: _CATALOG, OPENAPI_URL: _OPENAPI})
-    )
-    params = (await _contract(credential_blobs, authenticated_client))["parameters"]
-    assert params["route"]["gateway"]["status"] == "disabled"
-
-    row = next(
-        r for r in authenticated_client.get("/v1/models").json()["data"] if r["id"] == _MODEL
-    )
-    assert set(row["supported_parameters"]).isdisjoint({"route", "service_tier", "tool_choice_x"})
